+++ src/Speckle.Sdk/Serialisation/V2/Receive/ObjectLoader.cs
@@ -9,6 +9,7 @@ using Speckle.Sdk.Transports;
 
 namespace Speckle.Sdk.Serialisation.V2.Receive;
 
+public partial interface IObjectLoader : IDisposable;
 [GenerateAutoInterface]
 public sealed class ObjectLoader(
   ISqLiteJsonCacheManager sqLiteJsonCacheManager,
@@ -20,6 +21,8 @@ public sealed class ObjectLoader(
   private long _checkCache;
   private long _cached;
   private DeserializeProcessOptions _options = new(false);
+  [AutoInterfaceIgnore]
+  public void Dispose() => sqLiteJsonCacheManager.Dispose();
 
   public async Task<(string, IReadOnlyCollection<string>)> GetAndCache(
     string rootId,
